[#124] Rearranged module dependancy orders
