+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/ByteWritable.java
@@ -1,0 +1,7 @@
+package net.zscript.javaclient;
+
+import net.zscript.javaclient.ZscriptByteString.ZscriptByteStringBuilder;
+
+public interface ByteWritable {
+    ByteWritable writeTo(final ZscriptByteStringBuilder out);
+}
