+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/ZscriptByteString.java
@@ -1,4 +1,4 @@
-package net.zscript.javaclient.commandbuilder;
+package net.zscript.javaclient;
 
 import java.nio.charset.StandardCharsets;
