+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/addressing/ZscriptAddress.java
@@ -10,9 +10,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import net.zscript.javaclient.commandbuilder.ByteWritable;
-import net.zscript.javaclient.commandbuilder.ZscriptByteString;
-import net.zscript.javareceiver.tokenizer.TokenBuffer;
+import net.zscript.javaclient.ByteWritable;
+import net.zscript.javaclient.ZscriptByteString;
 import net.zscript.model.components.Zchars;
 import net.zscript.util.OptIterator;
