+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/commandPaths/Response.java
@@ -1,6 +1,6 @@
 package net.zscript.javaclient.commandPaths;
 
-import net.zscript.javaclient.commandbuilder.ZscriptByteString;
+import net.zscript.javaclient.ZscriptByteString;
 import net.zscript.model.components.Zchars;
 
 public class Response {
