+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/nodes/ZscriptNode.java
@@ -1,10 +1,7 @@
 package net.zscript.javaclient.nodes;
 
-import java.util.ArrayDeque;
-import java.util.Deque;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Queue;
 import java.util.function.Consumer;
 
 import net.zscript.javaclient.addressing.AddressedCommand;
