+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/sequence/AbstractLockCondition.java
@@ -1,7 +1,5 @@
 package net.zscript.javaclient.sequence;
 
-import java.util.HashSet;
-import java.util.Set;
 
 import net.zscript.javaclient.commandPaths.Command;
 import net.zscript.javaclient.commandPaths.CommandExecutionPath;
