--- clients/java-client-lib/client-command-api/src/main/java/net/zscript/javaclient/commandbuilder/ByteWritable.java
@@ -1,7 +1,0 @@
-package net.zscript.javaclient.commandbuilder;
-
-import net.zscript.javaclient.commandbuilder.ZscriptByteString.ZscriptByteStringBuilder;
-
-public interface ByteWritable {
-    ByteWritable writeTo(final ZscriptByteStringBuilder out);
-}
