+++ clients/java-client-lib/client-command-api/src/main/java/net/zscript/javaclient/commandbuilder/commandnodes/ZscriptCommandBuilder.java
@@ -9,8 +9,8 @@ import java.util.Map;
 
 import static java.util.stream.Collectors.joining;
 
-import net.zscript.javaclient.commandbuilder.ByteWritable;
-import net.zscript.javaclient.commandbuilder.ZscriptByteString.ZscriptByteStringBuilder;
+import net.zscript.javaclient.ByteWritable;
+import net.zscript.javaclient.ZscriptByteString.ZscriptByteStringBuilder;
 import net.zscript.javaclient.commandbuilder.ZscriptMissingFieldException;
 import net.zscript.javaclient.commandbuilder.ZscriptResponse;
 import net.zscript.model.components.Zchars;
