--- clients/java-client-lib/client-main/src/main/java/net/zscript/javaclient/connection/ExpressionSequence.java
@@ -1,37 +1,0 @@
-package net.zscript.javaclient.connection;
-
-import static net.zscript.javaclient.connection.ZscriptAddressPath.path;
-
-import java.util.List;
-
-abstract class ExpressionSequence<S extends ExpressionSequence<S>> {
-    /**
-     * Addresses in top-down hierarchical order.
-     */
-    List<ZscriptAddress> addressPath;
-    Integer              echoNumber;
-
-    ExpressionSequence() {
-    }
-
-    public S setEchoNumber(int echoNumber) {
-        this.echoNumber = echoNumber;
-        return (S) this;
-    }
-
-    public S appendSubAddress(ZscriptAddress subAddress) {
-        addressPath.add(subAddress);
-        return (S) this;
-    }
-
-    public S prependSubAddress(ZscriptAddress subAddress) {
-        addressPath.add(0, subAddress);
-        return (S) this;
-    }
-
-    public ZscriptAddressPath getAddressPath() {
-        return path(addressPath);
-    }
-
-    public abstract byte[] asZscriptBytes();
-}
