+++ clients/java-client-lib/client-main/src/main/java/net/zscript/javaclient/connectors/ZscriptConnectors.java
@@ -1,13 +1,8 @@
 package net.zscript.javaclient.connectors;
 
 import java.io.IOException;
-import java.util.Collection;
-import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Consumer;
 
-import net.zscript.javaclient.connection.ZscriptConnection;
 import net.zscript.javaclient.connectors.serial.SerialConnector;
 import net.zscript.javaclient.connectors.tcp.TcpConnector;
 
@@ -16,7 +11,7 @@ public class ZscriptConnectors {
     public static final SerialConnector SERIAL = new SerialConnector();
 
     public interface ZscriptConnector<T> {
-        ZscriptConnection connect(T underlyingImpl) throws IOException;
+        RawConnection connect(T underlyingImpl) throws IOException;
 
         /**
          * Starts a scan for possible connection, and posts each possible hit to the supplied Consumer.
