+++ clients/java-client-lib/client-main/src/main/java/net/zscript/javaclient/devices/ResponseSequenceCallback.java
