+++ clients/java-client-lib/pom.xml
@@ -16,9 +16,15 @@
     <modules>
         <module>client-command-api</module>
         <module>client-command-builders</module>
+        <module>client-main</module>
     </modules>
 
     <dependencies>
+        <dependency>
+            <groupId>net.zscript</groupId>
+            <artifactId>client-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <dependency>
             <groupId>net.zscript</groupId>
             <artifactId>zscript-model-standard-module-definitions</artifactId>
